updated id from int to long, added add method and get method

+++ src/main/java/bg/murval/maintenanceapi/models/Task.java
@@ -3,14 +3,14 @@ package bg.murval.maintenanceapi.models;
 import bg.murval.maintenanceapi.utils.Status;
 
 public class Task extends TaskRequest {
-    private final int id;
+    private final long id;
 
-    public Task(final int id, final String name, final String description, final Status status) {
+    public Task(final long id, final String name, final String description, final Status status) {
         super(name, description, status);
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
